+++ src/clj/l22/middleware.clj
@@ -13,7 +13,9 @@
     [buddy.auth.middleware :refer [wrap-authentication wrap-authorization]]
     [buddy.auth.accessrules :refer [restrict]]
     [buddy.auth :refer [authenticated?]]
-    [buddy.auth.backends.session :refer [session-backend]]))
+    [buddy.auth.backends.session :refer [session-backend]]
+    [ring.util.http-response :as response]
+    [taoensso.timbre :as timbre]))
 
 (defn wrap-internal-error [handler]
   (fn [req]
@@ -33,7 +35,6 @@
        {:status 403
         :title "Invalid anti-forgery token"})}))
 
-
 (defn wrap-formats [handler]
   (let [wrapped (-> handler wrap-params (wrap-format formats/instance))]
     (fn [request]
@@ -41,13 +42,22 @@
       ;; since they're not compatible with this middleware
       ((if (:websocket? request) handler wrapped) request))))
 
+;; (defn on-error [request response]
+;;   (error-page
+;;    {:status 403
+;;     :title (str "Access to " (:uri request) " is not authorized")}))
+
 (defn on-error [request response]
-  (error-page
-    {:status 403
-     :title (str "Access to " (:uri request) " is not authorized")}))
+  (response/found "/login"))
+
+(defn admin? [request]
+  (let [identity (get-in request [:session :identity] nil)]
+   (timbre/debug identity)
+   (boolean identity)))
 
 (defn wrap-restricted [handler]
-  (restrict handler {:handler authenticated?
+  (restrict handler {;;:handler authenticated?
+                     :handler admin?
                      :on-error on-error}))
 
 (defn wrap-auth [handler]